pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "audio-agent-md"
version = "0.1.0"
description = "Procedural, executable skills for real-time voice agents"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Avasis AI", email = "contact@avasis.ai" }
]
maintainers = [
    { name = "Avasis AI", email = "contact@avasis.ai" }
]
keywords = [
    "voice",
    "audio",
    "agent",
    "tts",
    "procedural",
    "voice-agent",
    "emotional-resonance",
    "real-time"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.urls]
Homepage = "https://github.com/avasis-ai/audio-agent-md"
Documentation = "https://github.com/avasis-ai/audio-agent-md#readme"
Repository = "https://github.com/avasis-ai/audio-agent-md.git"
Issues = "https://github.com/avasis-ai/audio-agent-md/issues"

[project.scripts]
audio-agent = "audio_agent_md.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/audio_agent_md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
multi_line_output = 3
